+++ NEWS.md
@@ -57,6 +57,7 @@ Documentation
 Enhancements:
 
 * Added the `output` argument to the `rbernoulli()` function to allow different output formats.
+* Change default of `sort_dag` in `sim_from_dag()` from `TRUE` to `FALSE`
 
 New Features: